chore(documentation): Added sub-navigation for layout module
