+++ packages/layout/examples/layout/layout.module.ts
@@ -13,7 +13,7 @@ import { LogoModule } from '@acpaas-ui/ngx-components/logo';
 import { CodeSnippetModule } from '@acpaas-ui/ngx-components/code-snippet';
 import { RouterModule } from '@angular/router';
 
-import { AUIDemoModalComponent } from './pages/demo/demo-modal.component';
+import { AUIDemoModalComponent } from './pages/modal/demo-modal.component';
 
 import { Pages } from './pages/index';
