+++ packages/layout/examples/layout/pages/cookie-consent/cookie-consent.page.html
@@ -1,0 +1,22 @@
+<h2 class="h4 u-margin-bottom">Cookie consent</h2>
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="cookie1"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="cookie2"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom">
+	<div class="m-alert">
+		<p>Note that since a cookie consent can take many forms, we did not provide a standard design for it. When dismissed a <code>.cc-invisible</code> class will be added. You can use this class to hide it from the user.</p>
+	</div>
+</div>
+
+<div class="u-margin-bottom">
+	<p>As an example, you can use the following SCSS code:</p>
+</div>
+
+<div class="u-margin-bottom-xx">
+	<aui-code-snippet [codeSnippet]="cookie3"></aui-code-snippet>
+</div>
