+++ packages/layout/examples/layout/pages/demo/demo.page.html
@@ -1,157 +1,15 @@
 <h1 class="h3 u-margin-bottom">Layout</h1>
 
-<!-- COOKIE CONSENT -->
-<h2 class="h4 u-margin-bottom">Cookie consent</h2>
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="cookie1"></aui-code-snippet>
-</div>
+<nav class="u-margin-bottom-lg">
+	<ul class="m-nav-tabs m-nav-tabs__layout">
+		<li><a [routerLink]="['cookie-consent']" [routerLinkActive]="['is-active']">Cookie consent</a></li>
+		<li><a [routerLink]="['footer']" [routerLinkActive]="['is-active']">Footer</a></li>
+		<li><a [routerLink]="['header']" [routerLinkActive]="['is-active']">Header</a></li>
+		<li><a [routerLink]="['hero']" [routerLinkActive]="['is-active']">Hero</a></li>
+		<li><a [routerLink]="['modal']" [routerLinkActive]="['is-active']">Modal</a></li>
+		<li><a [routerLink]="['pane']" [routerLinkActive]="['is-active']">Pane</a></li>
+		<li><a [routerLink]="['sidebar']" [routerLinkActive]="['is-active']">Sidebar</a></li>
+	</ul>
+</nav>
 
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="cookie2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<div class="m-alert">
-		<p>Note that since a cookie consent can take many forms, we did not provide a standard design for it. When dismissed a <code>.cc-invisible</code> class will be added. You can use this class to hide it from the user.</p>
-	</div>
-</div>
-
-<div class="u-margin-bottom">
-	<p>As an example, you can use the following SCSS code:</p>
-</div>
-
-<div class="u-margin-bottom-xx">
-	<aui-code-snippet [codeSnippet]="cookie3"></aui-code-snippet>
-</div>
-
-
-<!-- FOOTER -->
-<h2 class="h4 u-margin-bottom">Footer</h2>
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="footer1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="footer2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-xx">
-	<aui-footer isExtended="false">
-		<div class="u-margin-bottom u-margin-top">
-			<div auiFooterContent>
-				Footer content goes here
-			</div>
-			<div auiFooterBottom>
-				<aui-subfooter>
-					<aui-copyright domain="Digipolis"></aui-copyright>
-				</aui-subfooter>
-			</div>
-		</div>
-	</aui-footer>
-</div>
-
-
-<!-- HEADER -->
-<h2 class="h4 u-margin-bottom">Header</h2>
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="header1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="header2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-header>
-		<div auiHeaderContent>
-			<div auiHeaderLogo class="o-header__wrapper">
-				<aui-logo title="Example" src="https://robohash.org/acpaas-ui"></aui-logo>
-				<a class="o-header__link">Example</a>
-			</div>
-		</div>
-		<div auiHeaderMenuItem>
-			<a href="http://github.com/digipolisantwerp/acpaas-ui_angular" class="a-button a-button--navigation has-icon-left" target="_blank">
-				<i class="fa fa-github"></i>ACPaaS UI on GitHub
-			</a>
-		</div>
-	</aui-header>
-</div>
-
-<div class="u-margin-bottom-xx">
-	<div class="m-alert">
-		<p>As you may have noticed, the header on this page is a working example of the demo code.</p>
-	</div>
-</div>
-
-
-<!-- HERO -->
-<h2 class="h4 u-margin-bottom">Hero</h2>
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="hero1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="hero2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-xx">
-	<aui-hero>
-		<div auiHeroCard>
-			<h1>Your app</h1>
-			<a>Some link</a>
-		</div>
-		<div auiHeroCta>
-			<div class="buttons">
-				<a class="a-button">Home</a>
-				<a class="a-button">Another page</a>
-			</div>
-		</div>
-	</aui-hero>
-</div>
-
-
-<!-- MODAL -->
-<h2 class="h4 u-margin-bottom">Modal</h2>
-<div class="u-margin-bottom-xx">
-	<button class="a-button"
-		(click)="openModal()">
-		Test modal
-	</button>
-</div>
-
-
-<!-- PANE -->
-<h2 class="h4 u-margin-bottom">Pane</h2>
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="pane1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="pane2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="pane3"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-xx">
-	<button class="a-button" (click)="myPane.togglePane()">Toggle pane</button>
-	<button class="a-button" (click)="myPane.openPane()">Open pane</button>
-	<p class="u-margin-top-xs">The pane is <strong>{{ pane }}</strong>.</p>
-	<aui-pane #myPane
-		[side]="'left'"
-		[opened]="opened"
-		[backdrop]="backdrop"
-		(open)="onOpen()"
-		(close)="onClose()">
-		Pane content
-		<button class="a-button" (click)="myPane.closePane()">Close pane</button>
-	</aui-pane>
-</div>
-
-
-<!-- SIDEBAR -->
-<!-- <h2 class="h4 u-margin-bottom">Sidebar</h2>
-<div class="u-margin-bottom">
-	<aui-sidebar title="Demo sidebar" [items]="sidebarItems" open="true"></aui-sidebar>
-</div> -->
+<router-outlet></router-outlet>
