+++ packages/layout/examples/layout/pages/footer/footer.page.html
@@ -1,0 +1,23 @@
+<h2 class="h4 u-margin-bottom">Footer</h2>
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="footer1"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="footer2"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-xx">
+	<aui-footer isExtended="false">
+		<div class="u-margin-bottom u-margin-top">
+			<div auiFooterContent>
+				Footer content goes here
+			</div>
+			<div auiFooterBottom>
+				<aui-subfooter>
+					<aui-copyright domain="Digipolis"></aui-copyright>
+				</aui-subfooter>
+			</div>
+		</div>
+	</aui-footer>
+</div>
