+++ packages/layout/examples/layout/pages/header/header.page.html
@@ -1,0 +1,30 @@
+<h2 class="h4 u-margin-bottom">Header</h2>
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="header1"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="header2"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom">
+	<aui-header>
+		<div auiHeaderContent>
+			<div auiHeaderLogo class="o-header__wrapper">
+				<aui-logo title="Example" src="https://robohash.org/acpaas-ui"></aui-logo>
+				<a class="o-header__link">Example</a>
+			</div>
+		</div>
+		<div auiHeaderMenuItem>
+			<a href="http://github.com/digipolisantwerp/acpaas-ui_angular" class="a-button a-button--navigation has-icon-left" target="_blank">
+				<i class="fa fa-github"></i>ACPaaS UI on GitHub
+			</a>
+		</div>
+	</aui-header>
+</div>
+
+<div class="u-margin-bottom-xx">
+	<div class="m-alert">
+		<p>As you may have noticed, the header on this page is a working example of the demo code.</p>
+	</div>
+</div>
