+++ packages/layout/examples/layout/pages/hero/hero.page.html
@@ -1,0 +1,23 @@
+<h2 class="h4 u-margin-bottom">Hero</h2>
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="hero1"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="hero2"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-xx">
+	<aui-hero>
+		<div auiHeroCard>
+			<h1>Your app</h1>
+			<a>Some link</a>
+		</div>
+		<div auiHeroCta>
+			<div class="buttons">
+				<a class="a-button">Home</a>
+				<a class="a-button">Another page</a>
+			</div>
+		</div>
+	</aui-hero>
+</div>
