+++ packages/layout/examples/layout/pages/index.ts
@@ -1,5 +1,19 @@
 import { LayoutDemoPageComponent } from './demo/demo.page';
+import { LayoutCookieconsentDemoPageComponent } from './cookie-consent/cookie-consent.page';
+import { LayoutFooterDemoPageComponent } from './footer/footer.page';
+import { LayoutHeaderDemoPageComponent } from './header/header.page';
+import { LayoutHeroDemoPageComponent } from './hero/hero.page';
+import { LayoutModalDemoPageComponent } from './modal/modal.page';
+import { LayoutPaneDemoPageComponent } from './pane/pane.page';
+import { LayoutSidebarDemoPageComponent } from './sidebar/sidebar.page';
 
 export const Pages = [
 	LayoutDemoPageComponent,
+	LayoutCookieconsentDemoPageComponent,
+	LayoutFooterDemoPageComponent,
+	LayoutHeaderDemoPageComponent,
+	LayoutHeroDemoPageComponent,
+	LayoutModalDemoPageComponent,
+	LayoutPaneDemoPageComponent,
+	LayoutSidebarDemoPageComponent,
 ];
