+++ packages/layout/examples/layout/pages/modal/modal.page.html
@@ -1,0 +1,7 @@
+<h2 class="h4 u-margin-bottom">Modal</h2>
+<div class="u-margin-bottom-xx">
+	<button class="a-button"
+		(click)="openModal()">
+		Test modal
+	</button>
+</div>
