+++ packages/layout/examples/layout/pages/sidebar/sidebar.page.html
@@ -1,0 +1,4 @@
+<h2 class="h4 u-margin-bottom">Sidebar</h2>
+<div class="u-margin-bottom">
+	<aui-sidebar title="Demo sidebar" [items]="sidebarItems" open="true"></aui-sidebar>
+</div>
