+++ styleguide/index.html
@@ -7,7 +7,7 @@
 
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link rel="icon" type="image/x-icon" href="favicon.ico">
-	<link rel="stylesheet" href="https://cdn.antwerpen.be/acpaas_branding_scss/2.0.0/main.min.css">
+	<link rel="stylesheet" href="https://cdn.antwerpen.be/acpaas_branding_scss/2.1.0/main.min.css">
 
 	<link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/9.10.0/styles/atom-one-light.min.css">
 	<script src="//cdnjs.cloudflare.com/ajax/libs/highlight.js/9.10.0/highlight.min.js"></script>
