+++ styleguide/styles.scss
@@ -182,3 +182,7 @@ pre {
 		padding: $spacer / 2;
 	}
 }
+
+.m-nav-tabs__layout {
+	box-shadow: 0 -1px 0 $border-color inset;
+}
